+++ instrumentation/spring/spring-ws-2.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/spring/ws/v2_0/SpringWsTest.java
@@ -50,7 +50,7 @@ class SpringWsTest extends AbstractHttpServerUsingTest<ConfigurableApplicationCo
   private static final Jaxb2Marshaller marshaller = new Jaxb2Marshaller();
 
   private static final boolean CONTROLLER_TELEMETRY =
-      Boolean.getBoolean("otel.instrumentation.common.experimental.controller-telemetry.enabled");
+      Boolean.getBoolean("otel.instrumentation.common.controller-telemetry.enabled");
 
   @BeforeAll
   void setup() {